+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG.md
 
+## v0.3.0 (2019-10-08)
+
+* Moved to github.com/ace-teknologi/memzy
+* Added the Memzy interface
+* Added the Iter interface
+* Added an in-memory implementation of memzy
+
 ## v0.2.0 (2019-04-26)
 
 * Added GetItem functionality